Its possible that the people who are closing the connection as soon as we send are just trying to ruin the torrent or something. We actually serve people who do not destroy the connection from time to time...
Really not sure why though.

NEED TO TEST OUT UPLOADING TO OTHER BITRAPTOR PEERS AND STUFF

Oh, and fixed seeder (I bet) because now we advertise bitfield (if we have receivedPieces) to any new peer. Before, the other client would not have seen anything and assumed it was just a leecher with 0 things.